Make services more relative
Currently the objects that implement services have to instance variables
that contain a reference to the connection and the absolute path. For
example the service that manages virtual machine '122' has the following
instance variables:

  @connection = ...
  @path = 'vms/123'

This works fine, but it makes it complicates replacement of the
connection object, as all the services hold a reference to it.

To improve that this patch changes the implementation of the services so
that will have a reference to the parent service, and a relative path:

  @parent = ...
  @path = '123'

The parent of a service will usually be another service. For example,
the parent of the service that manages virtual machine '123' will be the
service that manages the collection of virtual machines. For root of the
tree of services (the 'SystemService') the parent will be the
connection.

Change-Id: I2bdadaed8cd4e0954fc6a3600a01d2507e6db4da
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9f4174538bc761923517ab8f21165a8d3fd8df55)