Memoize references to services without parameters
Currently every time that a service is located a new Ruby object is
created. For example, this is the implementation of the method that
returns the reference to the service that manages the collection of
virtual machines:

  def vms_service
    VmsService.new(self, 'vms')
  end

Creating these identical objects is a waste of resources, specially for
services that are going to be used many times. To avoid that waste this
patch changes the SDK so that those objects will be created only once:

  def vms_service
    @vms_service ||= VmsService.new(self, 'vms')
  end

Note that this will not affect the services that manage a single object.
For example, the object corresponding to the service that manages a
single virtual machine will be created every time that the
'vms_service.vm_service(id)' is called.

Change-Id: I7c51fce34635565ab6178681484abf7320f5364a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>