Fix URL building
A recent patch introduced a regression in the logic that builds URLs, as
a result services with more than one segment in the path don't work
correctly. For example, the following code:

  connection.system_service()
    .vms_service(123)
    .disk_attachments_service()
    .attachment_service(456)

Does not build the expected path:

  vms/123/diskattachments/456

This patch fixes the regression and adds a test to make sure that it
won't be introduced again.

Change-Id: I8ae4585c056337099a1dd302ea7eed36a98ebe63
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>