Improve log for response codes
Currently the log message generated for response codes only contains the
code itself. For example:

  Received response code '200'.

This is fine when requests are sent in synchronous mode, as it is clear
what is the corresponding response. But when using asynchronous mode it
is difficult to correlate the request with the response. To improve that
this patch changes the SDK so that the log messages will contain the
response code and also some additionl details about the request:

  Received response code 200 for GET request to URL 'https://...'.

Change-Id: I51f8a2406edaf977fc82c94f43c848848519084a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>