Add specific exceptions
Currently the SDK raises an exception of 'Error' type for any error it
detects. This complicates things for applications that need to handle
specific errors, as they need to parse the error message in order to
decide what to do. To improve that this patch adds net exception
classes, extending 'Error', for some frequently needed kinds of errors:

  AuthError - Raised when authentication or authorization fail.

  ConnectionError - Raised when the name of the server can’t be
  resolved, and when the server is down or unreachable.

  NotFoundError - Raised when the requested object doesn’t exist.

  TimeoutError - Raised when an operation times out.

Existing applications don't need to be modified, unless they want to
take advantage of the new exceptions.

Change-Id: Icc0129f2cd15ddc7a6e2d6d3981ccfc56601c152
Bug-url: https://bugzilla.redhat.com/1465328
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>