Support OAuth and OpenID
Currently the code that parses the error reponses from the SSO service
assumes that it is using plain OAuth. But the service may also use
OpenID, which uses a different set of attributes for error responses.
This patch changes the SDK so that it handlers correctly both kind of
responses.

Change-Id: I444dd4b3e83b1fe27c9f4bd85b5d28dd09b20d6c
Bug-Url: https://bugzilla.redhat.com/1468461
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>