Check service method parameters
Currently when a method is invoked with a parameter that is not
supported it is just silently ignored. This patch changes the SDK so
that it will check the parameters, and generate an exception when
unsupported parameters are used. For example, the following program
tries to use the 'cloud_init' parameter that doesn't exist (the correct
name is 'use_cloud_init'):

   vm_service.start(
     cloud_init: true,
     vm: {
       ...
     }
   )

Before this change that started the virtual machine, ignoring the wrong
parameter, and with cloud-init disabled. After this change the program
will fail with the following exception:

  .../lib/ovirtsdk4/service.rb:427:in `check_bad_opts': The option
  'cloud_init' isn't supported. The supported options are 'async',
  'filter', 'pause', 'use_cloud_init', 'use_sysprep' and 'vm'.
  (OvirtSDK4::Error)
    from .../lib/ovirtsdk4/service.rb:337:in `internal_action'
    from .../lib/ovirtsdk4/services.rb:26232:in `start'
    from start_vm.rb:56:in `<main>'

Change-Id: Iec8d7443896e79dfd7061ef370430f57f0382116
Bug-Url: https://bugzilla.redhat.com/1378113
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>