Add support for multiple threads
Currently the SDK doesn't support multiple threads using the same
connection simultaneously. If the application tries to do that the
result isn't predictable, but usually a segmentation fault. The reason
for that is that the underlying library, libcurl, doesn't support
multiple threads either. But we have situations where we need it. For
example, the ManageIQ events collector worker has two threads, and both
may use the same connection simultaneously. In order to support those
usages this patch adds a mechanism to allow access by multiple threads.
The mechanism is quite simple: a mutex that only allows one thread in
the 'Connection.send' and 'Connection.wait' methods.

Change-Id: I5b863fa7397aeb03e0c0e464c0a500f466c3a64c
Bug-Url: https://bugzilla.redhat.com/1496846
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>