Improve the example that shows how to add NIC to a VM
The example that shows how to add a NIC to a VM doesn't take into
account that there may be multiple vNIC profiles with the same name.
This patch updates the example so that it locates the vNIC profile
navigating from the VM to the cluster, then to the networks, and finally
to the vNIC profiles.

Change-Id: I20b64db9bbaf6e4f06cce3aa8ad889d5e8716352
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>