Don't generate code for specialized signatures
The SDK isn't yet prepared to generate code for the specialized
signatures of some methods. As a result the metadata used to define the
types of the parameters can be duplicated and cause warnings when
loading the code of the SDK. To avoid that this patch changes the code
generator so that it ignores these specialized signatures.

Change-Id: I61a58ad61e867b91d16613c8021c0688e774e37d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>