Check types of HttpClient constructor parameters
This patch changes the constructor of the HttpClient class so that it
checks the types of the parameters. This checking may abort the
initialization of the object, leaving some of the fields un-initialized,
which causes segmentation faults when the memory is released. So to
avoid that we also need to make sure that all the fields are initialized
to their zero values when the memory is initially allocated.

Change-Id: I4d28586a82a2f5f47438a2b89cbb897aeaaea8ce
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 8cef59c7cafa46242fef1e77a8787a6704e68288)