Revert use of `Time` instead of `DateTime`
The previous patch replaced the usage of the `DateTime` class with
`Time`, because Rubocop complains about it. But we do need `DateTime`
because we use the `xmlschema` method that isn't available in the `Time`
class. So this patch reverts the previous change, and modifies the
Rubocop configuration so that this check is disabled.

Change-Id: I2b14b9b12443485cf06326c10d663fa2dbd3973e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 15b29d263dbf43fe7121bbe24a3428ab35b29e28)