Require `redhat-rpm-config` for building
When building the SDK in Fedora, CentOS or RHEL using the version of
Ruby included in the distribution, it is also necessary to install the
`redhat-rpm-config` package. The reason is that the Makefile generated by
those versions of Ruby requires the `/usr/lib/rpm/redhat/redhat-hardened-cc1`
file, if it isn't present the build fails. That file is provided by the
`redhat-rpm-config` package, but not required by the `ruby-devel` package.
To avoid that issue this patch documents the need for `redhat-rpm-config`
in the build instructions.

Change-Id: Ie2e274706f64bb82f1dcc5bd20114b7b05f3f9c2
Bug-Url: https://bugzilla.redhat.com/1505427
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>