Don't send 'Expect: 100-continue'
The 'Expect: 100-continue' is an HTTP mechanism intended to allow
servers to reject requests inmediately after receiving the HTTP header,
without having to receive the potentially large request body. But the
oVirt server doesn't honour this header. As a result the underlying
libcurl library sends the header, and waits up to 1 second for the
response from the server. This adds 1 second to every POST and PUT
requests, without any benefit. To avoid that this patch changes the SDK
so that it explicitly disables this mechanism (for PUT requests, it was
already disabled for POST requests).

Change-Id: I2760f0831688eff941b9646441e95b6b8ab08099
Bug-Url: https://bugzilla.redhat.com/1509910
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>