Don't include sensible data in `inspect`
Currently the `inspect` method of the SDK objects may return sensible
data, in particular the result of the `inspect` method contains the user
name and password. This can be dangerous if the result of that `inspect`
method is written to log files. To avoid that this patch changes the SDK
objects so that they implement the `inspect` and `to_s` methods in a way
that does not contains sensible data.

Change-Id: I9be06822d8522789a79f63ebe44c65ad793058e0
Bug-Url: https://bugzilla.redhat.com/1513620
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>