Limit the number of requests sent to libcurl
The `libcurl` library has different options to set the number of
connections, added in different versions of the library. The option that
is more convenient for the SDK is `CURLMOPT_MAX_HOST_CONNECTIONS`, as it
limits the number of connections to one specific host. The SDK sets this
option to the value of the `connections` parameter. But this option was
added in version 7.30.0 of `libcurl`, which isn't available in CentOS 7,
for example. That means that in CentOS the `connections` setting isn't
honoured by the SDK, and that may result in an unlimited number of
simultaneous connections, whicy may in turn exhaust the resources of the
oVirt web server.

To improve that, this patch changes the SDK so that it controls
explicitly the number of requests sent to `libcurl` using a queue and a
calculated limit.

For versions of libcurl that support fine tuning of the number of
connections the limit will be calculated multiplying the maximum number
of connections by the maximum pipeline length.

For versions of `libcurl` that don't support that fine tuning (older
than 7.30.0) the limit will be much the maximum number of connections.

In addition, this patch also modifies the SDK so that it sends warning
messages to the log if the number of connections and the pipeline length
requested by the user can't be honoured.

Change-Id: Iee50595dde775bd377007fee4fee495e8f2fc950
Bug-Url: https://bugzilla.redhat.com/1525302
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>