Fix handling of the `all_content` parameter
A previous patch had the side effect of changing how query parameters
are stored in memory: they used to be stored as strings, and they are
now stored as symbols. This has a negative side effect in the handling
of the `all_content` parameter, because we check for the presence of the
`all_content` string instead of for the presence of the `:all_content`
symbol. The result is that the `All-Content` header isn't added. This
patch fixes that issue.

Change-Id: I213fbbaebd41ef976b52dd7516addff2ee22a546
Bug-Url: https://bugzilla.redhat.com/1525555
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>