Update versions of dependencies
This patch updates the versions of the dependencies to the latest
available. In particular it updates the version of the `yard` gem to at
least `0.9.12` to avoid a security vulnerability that was fixed in
`0.9.11`.

Note that the `yard` gem is used only in the development environment, to
generate the documentation, so this vulnerability doesn't have any
effect on applications using the SDK.

Change-Id: I9bfad6f88fee01416e1af33f2a1dddb99d65fcd9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>